+++ src/main/java/org/realityforge/graphql/domgen/ExceptingDataFetcher.java
@@ -6,26 +6,6 @@ import javax.annotation.Nonnull;
 
 @FunctionalInterface
 public interface ExceptingDataFetcher<T>
+  extends DataFetcher<T>
 {
-  T get( @Nonnull DataFetchingEnvironment environment )
-    throws Exception;
-
-  default DataFetcher toDataFetcher()
-  {
-    return environment ->
-    {
-      try
-      {
-        return get( environment );
-      }
-      catch ( final RuntimeException e )
-      {
-        throw e;
-      }
-      catch ( final Exception e )
-      {
-        throw new WrapperRuntimeException( e );
-      }
-    };
-  }
 }